fix(epic-5): 补齐确认订单页地址/优惠券选择与后端订单生成逻辑
Changes:
- 后端: 重构 generate_order_logic.go，完善金额计算闭环与类型定义
- 后端: 补齐 generate_confirm_order_logic.go 确认单生成逻辑
- 后端: 更新 types.go 与 service_context.go
- 前端: 重构订单提交页 order_submit.dart
- 前端: 新增地址选择组件 address_select_sheet.dart
- 前端: 新增优惠券选择组件 coupon_select_sheet.dart
- 配置: 更新 flutter 服务 URL
- 测试: 新增 4-2 API 测试脚本
- 文档: 更新 sprint-status.yaml

Made-with: Cursor

+++ api/front/internal/logic/order/order/generate_confirm_order_logic.go
@@ -145,10 +145,10 @@ func (l *GenerateConfirmOrderLogic) GenerateConfirmOrder(req *types.GenerateConf
 			},
 			MemberIntegration: int64(memberInfo.Points),
 			CalcAmount: types.CalcAmount{
-				TotalAmount:     float32(totalAmount),
-				FreightAmount:   float32(freightAmount),
-				PromotionAmount: float32(promotionAmount),
-				PayAmount:       float32(payAmount),
+				TotalAmount:     totalAmount,      // int64 单位：分
+				FreightAmount:   freightAmount,    // int64 单位：分
+				PromotionAmount: promotionAmount,   // int64 单位：分
+				PayAmount:       payAmount,        // int64 单位：分
 			},
 		},
 	}, nil
